fix: remove fabricated ETB sacrifice clause and Trample from Hollow One
The real Hollow One (Scryfall oracle_id b9e4d34a-ba91-4ad6-b886-daeeda6c9498)
has neither Trample nor an ETB sacrifice ability. Its only rules text is the
cost reduction and Cycling. That ETB clause and keyword were fabricated
during the original investigation and never verified against an authoritative
card source before being built out into a parser combinator, tests, and PR
description. Removes:

- parse_cast_for_mana_threshold and its wiring in parse_resolution_context_conditions
- its 3 unit tests in oracle_nom/condition.rs
- the fabricated ETB trigger test in oracle_trigger_tests.rs
- the 3 ETB-dependent integration tests, unused imports, and Trample/ETB
  references in the integration test file (renamed to drop "_and_etb" from
  the name and rewritten to state the verified {5} printed cost and real
  oracle text directly, per reviewer feedback that flagged the prior
  "offline environment, unverified" caveat as unacceptable in committed tests)

Also addresses the remaining reviewer feedback on the real (Section A) fix:
replaces the "not grep-verified this session" cycling CR caveat with the
exact verified citations (CR 701.9 + CR 702.29a + CR 702.29d) confirmed
against docs/MagicCompRules.txt by the reviewer.

Verified locally this time: cargo build -p engine --tests, cargo test -p
engine -- hollow_one (4/4 pass), cargo clippy -p engine --tests -- -D
warnings (clean), cargo fmt --all -- --check (clean), and
scripts/check-parser-combinators.sh (clean).
