+++ crates/engine/tests/integration/main.rs
@@ -128,7 +128,7 @@ mod harrow_regression;
 mod heist_production_path_handoff;
 mod hellkite_tyrant_steal_artifacts_2906;
 mod heroic_defiance_recipient_color_4590;
-mod hollow_one_cost_reduction_and_etb;
+mod hollow_one_cost_reduction;
 mod hunters_insight_combat_draw;
 mod inevitable_betrayal_no_mana_cost;
 mod integration_adventure;
